Documentação.

+++ src/main/java/com/caio/desafio/backend/itau/dto/TransacaoDto.java
@@ -1,6 +1,7 @@
 package com.caio.desafio.backend.itau.dto;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
-public record TransacaoDto(Double valor, OffsetDateTime dataHora) {
+public record TransacaoDto(BigDecimal valor, OffsetDateTime dataHora) {
 }
